refactor(terminal): split main.cpp into modules (move-only)
Closes #111 (child of terminal epic #73). Pure code move, no behaviour
change: main.cpp went 2979 -> 241 lines.

The single anonymous namespace is now a named `zwt` namespace carved across
header+cpp module pairs under utils/zwave-terminal/:
- constants.hpp        — compile-time constants
- activity.{hpp,cpp}   — shared UI state (log, dongle status, DaemonError) + logLine
- format.{hpp,cpp}     — status/CC name tables + wire-value renderings
- prompts.{hpp,cpp}    — bottom-row modal input helpers
- render.{hpp,cpp}     — draw() + runActionMenu() + MenuItem
- policy_blob.{hpp,cpp} — the deliberately-duplicated PolicyRegister BLOB codec
- signal_handlers.{hpp,cpp} — D-Bus signal subscriptions
- handlers.{hpp,cpp}   — interactive action handlers
main.cpp keeps only run() (the input loop) behind a thin global main().

Keeps the free-functions-in-namespace style (it's a utils/ tool, not the
daemon). Every TU is include-what-you-use clean under the strict
`misc-include-cleaner` + `WarningsAsErrors: '*'` config; TU-local helpers
sit in nested anonymous namespaces. sdbus-c++'s non-self-contained
Message.h (uses std::copy_n) is handled by an `<algorithm> // IWYU pragma:
keep` in the headers that pull it.

Builds clean under gnu + llvm; clang-tidy clean; ctest 257/257 unchanged.
